requirements.txt
Flask==2.3.2
pandas==2.2.3
anndata==0.11.3
scipy==1.15.2
numpy==2.1.3
scikit-learn==1.5.2
scanpy==1.11.0
kafka-python==2.0.4
joblib==1.4.2
flask-cors==3.0.10
